update
- 쓸모없는 중간 테이블 코드 삭제
- 프론트에 태그 기능 추가
