+++ src/main/java/com/example/dbdesign_project/playlist/PlaylistDAO.java
@@ -1,18 +1,17 @@
 package com.example.dbdesign_project.playlist;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Isolation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Repository
 public class PlaylistDAO {
-    private final JdbcTemplate jdbcTemplate;
-
-    public PlaylistDAO(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
     // 특정 사용자에 대한 모든 재생목록 조회
     public List<Playlist> getPlaylistsByUserId(int userId) {
         String sql = "SELECT * FROM Playlist WHERE userId = ?";
@@ -32,28 +31,22 @@ public class PlaylistDAO {
         return jdbcTemplate.update(sql, playlist.getUserId(), playlist.getListName());
     }
 
-    // 재생목록 이름 갱신
+    //이름 변경
     public int updatePlaylistName(int listId, String newName) {
         String sql = "UPDATE Playlist SET listName = ? WHERE listId = ?";
         return jdbcTemplate.update(sql, newName, listId);
     }
 
-    // 재생목록 삭제
+    // 삭제
     public int deletePlaylist(int listId) {
         String sql = "DELETE FROM Playlist WHERE listId = ?";
         return jdbcTemplate.update(sql, listId);
     }
 
-    // 특정 재생목록 이름 조회
+    //조회
     public String getPlaylistNameById(int listId) {
         String sql = "SELECT listName FROM Playlist WHERE listId = ?";
         return jdbcTemplate.queryForObject(sql, new Object[]{listId}, String.class);
     }
 
-    // 특정 재생목록 존재 여부 확인
-    public boolean existsById(int playlistId) {
-        String sql = "SELECT COUNT(*) FROM Playlist WHERE listId = ?";
-        Integer count = jdbcTemplate.queryForObject(sql, new Object[]{playlistId}, Integer.class);
-        return count != null && count > 0;
-    }
 }
