+++ src/main/resources/templates/song.html
@@ -132,6 +132,62 @@
             font-size: 0.8rem;
             margin-left: 10px;
         }
+        /* 태그 삭제 버튼 스타일 */
+        .tag-delete-btn {
+            width: 15px;
+            height: 15px;
+            font-size: 0.7rem;
+            background-color: #ff4d4d;
+            color: #fff;
+            border: none;
+            border-radius: 50%;
+            cursor: pointer;
+        }
+
+        .tag-delete-btn:hover {
+            background-color: #e60000;
+        }
+
+        /* 태그 추가 입력창 및 버튼 스타일 */
+        .tag-input {
+            width: 150px;
+            padding: 0.3rem;
+            font-size: 1rem;
+            border: none;
+            border-radius: 5px;
+            margin-right: 5px;
+        }
+
+        .add-tag-btn {
+            width: auto;
+            padding: 0.3rem 0.5rem;
+            font-size: 1rem;
+            background-color: #11998e;
+            color: #fff;
+            border: none;
+            border-radius: 5px;
+            cursor: pointer;
+        }
+
+        .add-tag-btn:hover {
+            background-color: #0d7f70;
+        }
+
+        /* 노래 삭제 버튼 스타일 */
+        .delete-btn {
+            width: auto;
+            padding: 0.3rem 0.5rem;
+            font-size: 1rem;
+            background: linear-gradient(135deg, #341f97 0%, #11998e 100%);
+            color: #fff;
+            border: none;
+            border-radius: 5px;
+            cursor: pointer;
+        }
+
+        .delete-btn:hover {
+            background-color: #0d7f70;
+        }
     </style>
 </head>
 
@@ -193,16 +249,40 @@
     <section class="right-section">
         <h2>노래 목록</h2>
         <ul>
-            <li th:each="song : ${songs}">
-                <span class="song-info" th:text="${song.artist} + ' - ' + ${song.songName} + ' (' + ${song.releasedYear} + ')'">노래 정보</span>
-                <form action="/playlistsong/remove" method="post" style="display:inline; margin-left:10px;">
-                    <input type="hidden" name="listId" th:value="${selectedListId}">
-                    <input type="hidden" name="songId" th:value="${song.songId}">
-                    <button type="submit" class="delete-btn">노래 삭제</button>
-                </form>
+            <li th:each="song : ${songs}" style="margin-bottom: 20px;">
+                <!-- 노래 정보와 태그 삭제 버튼 -->
+                <div style="display: flex; align-items: center;">
+                    <span class="song-info" th:text="${song.artist} + ' - ' + ${song.songName} + ' (' + ${song.releasedYear} + ')'">노래 정보</span>
+                    <span> | 태그: </span>
+                    <span th:each="tag : ${song.tags}" style="display: flex; align-items: center; margin-left: 10px;">
+                    <form action="/songtag/remove" method="post" style="display: inline; margin-right: 5px;">
+                        <input type="hidden" name="songId" th:value="${song.songId}">
+                        <input type="hidden" name="tagName" th:value="${tag}">
+                        <button type="submit" class="tag-delete-btn">x</button>
+                    </form>
+                    <span th:text="${tag}" style="margin-right: 10px;">태그 이름</span>
+                </span>
+                </div>
+
+                <!-- 태그 추가와 노래 삭제 버튼 -->
+                <div style="display: flex; margin-top: 10px;">
+                    <form action="/songtag/add" method="post" style="margin-right: 10px;">
+                        <input type="hidden" name="songId" th:value="${song.songId}">
+                        <input type="text" name="tagName" placeholder="태그 추가" class="tag-input" required>
+                        <button type="submit" class="add-tag-btn">+</button>
+                    </form>
+                    <form action="/songs/delete" method="post">
+                        <input type="hidden" name="listId" th:value="${selectedListId}">
+                        <input type="hidden" name="songId" th:value="${song.songId}">
+                        <button type="submit" class="delete-btn">노래 삭제</button>
+                    </form>
+                </div>
             </li>
         </ul>
     </section>
+
+
+
 </div>
 </body>
